+++ NEWS.md
@@ -2,5 +2,6 @@
 
 * Added BayesRates to Import list.
 * Imported 'n_draw' function from BayesRates.
+* Storing random seeds