+++ NEWS.md
@@ -1,3 +1,8 @@
+
+# BayesProj 0.1.2
+
+* Corrected error in use of @noRd in documentation
+
 # BayesProj 0.1.1
 
 * Added BayesRates to Import list.